Readability & Conciseness

+++ src/config.jl
@@ -1,7 +1,7 @@
 
 struct Config
-    k::Int64
-    step::Int64
-    threshold::Int64
-    nthreads::Int64
+    k::Int
+    step::Int
+    threshold::Int
+    nthreads::Int
 end
